requirements.txt
# Cattle Breed Recognition - Python Dependencies
# 100% Free and Open Source

# ============================================
# Core ML Framework
# ============================================
tensorflow>=2.12.0
keras>=2.12.0

# ============================================
# Computer Vision
# ============================================
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# ============================================
# Data Processing
# ============================================
numpy>=1.23.0
pandas>=2.0.0
scipy>=1.10.0

# ============================================
# Machine Learning Utilities
# ============================================
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================
# YOLO Support
# ============================================
# For YOLO model training and export
ultralytics>=8.0.0

# ============================================
# Data Collection
# ============================================
requests>=2.31.0
beautifulsoup4>=4.12.0
kaggle>=1.5.0

# ============================================
# Image Annotation
# ============================================
# LabelImg for bounding box annotation
# Install separately: pip install labelImg

# ============================================
# Model Export
# ============================================
onnx>=1.14.0
onnx-tf>=1.10.0

# ============================================
# Mobile App Support
# ============================================
# TFLite support is included in TensorFlow

# ============================================
# Web Dashboard (Expert Interface)
# ============================================
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# ============================================
# Testing
# ============================================
pytest>=7.3.0
pytest-cov>=4.1.0

# ============================================
# Utilities
# ============================================
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# ============================================
# Google Colab Specific
# ============================================
# These are pre-installed in Colab
# google-colab
# google-auth
# google-drive-downloader

# ============================================
# Optional: For Local Development
# ============================================
# jupyter>=1.0.0
# ipykernel>=6.23.0
# notebook>=6.5.0
